// resources/scripts/interfaces/WorkDisabilityTypeInterface.ts
export interface WorkDisabilityTypeInterface {
  workDisabilityTypeId: number
  workDisabilityTypeName: string
  workDisabilityTypeDescription: string
  workDisabilityTypeSlug: string
  workDisabilityTypeActive: number
  workDisabilityTypeCreatedAt?: string
  workDisabilityTypeUpdatedAt?: string
  deletedAt?: string
}