// resources/scripts/interfaces/WorkDisabilityPeriodExpenseInterface.ts
export interface WorkDisabilityPeriodExpenseInterface {
  [key: string]: any
  workDisabilityPeriodExpenseId: number
  workDisabilityPeriodExpenseFile: string
  workDisabilityPeriodExpenseAmount: number
  workDisabilityPeriodId: number
  workDisabilityPeriodExpenseCreatedAt?: string
  workDisabilityPeriodExpenseUpdatedAt?: string
  deletedAt?: string
}